requirement.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
django-qr-code==3.1.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pydantic==1.9.2
pyparsing==3.0.9
pytz==2022.1
qrcode==7.3.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
segno==1.5.2
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
